dashboard/requirements.txt
flask_socketio
flask_cors
wifi
flask_restful
pyserial
adafruit-circuitpython-adxl34x
adafruit-blinka
pynmea2
termcolor
Flask-SQLAlchemy==2.1.0
flask_marshmallow
rpi-backlight